[Exercise 2.5] Course module
