+++ course_info/src/App.jsx
@@ -1,31 +1,4 @@
-const Header = ({ course }) => <h1>{course}</h1>;
-
-const Total = ({ parts }) => {
-  return (
-    <p>
-      Number of exercises {parts.reduce((acc, part) => acc + part.exercises, 0)}
-    </p>
-  );
-};
-
-const Part = ({ part }) => (
-  <p>
-    {part.name} {part.exercises}
-  </p>
-);
-
-const Content = ({ parts }) =>
-  parts.map((part) => <Part key={part.id} part={part} />);
-
-const Course = ({ courses }) => {
-  return courses.map((course) => (
-    <div key={course.id}>
-      <Header course={course.name} />
-      <Content parts={course.parts} />
-      <Total parts={course.parts} />
-    </div>
-  ));
-};
+import Course from "./components/Course";
 
 const App = () => {
   const courses = [
