+++ course_info/src/components/Course.jsx
@@ -1,0 +1,31 @@
+const Header = ({ course }) => <h1>{course}</h1>;
+
+const Total = ({ parts }) => {
+  return (
+    <p>
+      Number of exercises {parts.reduce((acc, part) => acc + part.exercises, 0)}
+    </p>
+  );
+};
+
+const Part = ({ part }) => (
+  <p>
+    {part.name} {part.exercises}
+  </p>
+);
+
+const Content = ({ parts }) =>
+  parts.map((part) => <Part key={part.id} part={part} />);
+
+const Course = ({ courses }) => {
+  // desctructure the props object, so asking for courses array as argument is fine
+  return courses.map((course) => (
+    <div key={course.id}>
+      <Header course={course.name} />
+      <Content parts={course.parts} />
+      <Total parts={course.parts} />
+    </div>
+  ));
+};
+
+export default Course;
